+++ activerecord/lib/active_record/timestamp.rb
@@ -48,7 +48,7 @@ module ActiveRecord
         current_time = current_time_from_proper_timezone
 
         all_timestamp_attributes.each do |column|
-          if respond_to?(column) && respond_to?("#{column}=") && self.send(column).nil?
+          if attributes.has_key?(column.to_s) && self.send(column).nil?
             write_attribute(column.to_s, current_time)
           end
         end